+++ src/opCode/chunk.cpp
@@ -32,7 +32,7 @@ void writeChunk(Chunk* chunk, uint8_t byte, int line) {
     chunk->count++;
 }
 
-int addConstants(Chunk* chunk, Value value) {
+size_t addConstants(Chunk* chunk, Value value) {
     writeValueArray(&chunk->constants, value);
     return chunk->constants.count - 1;
 }
