+++ src/opCode/chunk.hpp
@@ -46,4 +46,4 @@ void initChunk(Chunk* chunk);
 void freeChunk(Chunk* chunk);
 void writeChunk(Chunk* chunk, uint8_t byte, int line);
 
-int addConstants(Chunk* chunk, Value value);
+size_t addConstants(Chunk* chunk, Value value);
